Updated board udev rules to work with Ubuntu 18.04
For some reason, udev doesn't fire with the same environment when adding
or removing a board as in Ubuntu 16.04. These new rules seem to reliably
match on both. A side-effect of this is that Digilent JTAG devices now
also appear as a tty ('dftdrvdtch' is not run).

Also updated the names of rules files and JTAG symlinks.

Closes #17.