Fixed issue with VC709 tty enumeration using new udev rules
The VC709's JTAG tty device was overriding the CP210x UART as it was matching
through the Zybo's UART rules. Added a check to these to ensure serial number
is likely to actually be a Zybo (boards with a CP210x UART must have the first
8 digits of the JTAG serial set to '0' as the UART chip only allows 4 digits to
be used). This should work fine, as long as we never get a Zybo with a serial
that matches '00000000????'.