More reliably kill board server Docker containers on board detach
Previously the container would be left running if the board host
couldn't contact the relay for whatever reason (network issues, relay
not running, etc.), but we actually want the container to be killed no
matter what, then the relay can still tidy up if it was a temporary
network disruption.

This could maybe have led to some of the behaviour observed in #24.