Filtered out unwanted/erroneous firings of the board detach script
Sometimes (e.g. when the Digilent utilities remove unneeded tty drivers)
boards are 'detached' by udev when they are actually still running
totally fine. The board detach script will now not run if all the device
symlinks in /dev/vlab/ are present and resolve ok.

The detach script also now does not run if the board's Docker container
is not running (as this just caused an error further along anyway).

Closes #26