Reset the FPGA configuration of marked boards when a user connects
Previously this just happened on disconnect, but if a board is marked as
wanting to be reset after use, then it's better to do it on connect too
so that it's *definitely* clean when a user's session starts (in case
the previous session disconnected badly and left it configured somehow).

First step of solving #28.