Added idea of in-use/available boards, as well as locked/unlocked boards
Now a board will remain as 'in use' as long as a user is connected to
it, with 'shell.py' pinging the Redis database every 10 seconds while
screen is running. Board 'sessions' are created when a user connects,
and are ended when the user cleanly disconnects, when another user takes
a board, or when 'checkboards.py' detects that a shell has not pinged
the server for 30 seconds.

A board's FPGA configuration will be reset (if appropriate) when stale
sessions are cleaned up by 'checkboards.py', and its board server
Docker container will also be restarted to ensure the user is actually
disconnected.

'shell.py' will also now actively unlock the user's board when the lock
expires, so the system no longer just relies on 'checkboards.py'
enforcing and cleaning up locks.

Closes #13
Closes #28